requirements.txt
torch==2.3.1
torchaudio==2.3.1
numpy==1.26.4
scipy==1.11.4
scikit-learn==1.4.2
mne==1.6.1
matplotlib==3.8.4
umap-learn==0.5.6
einops==0.7.0
tqdm==4.66.4
pyyaml==6.0.1
